move structdef to ContractDataDef ns
