+++ test/test.bctklib/ContractDataDefParserTest.cs
@@ -5,9 +5,10 @@ using System.Linq;
 using FluentAssertions;
 using OneOf;
 
-using FieldType = OneOf.OneOf<Neo.SmartContract.ContractParameterType, Neo.BlockchainToolkit.StructDef>;
+using FieldType = OneOf.OneOf<Neo.SmartContract.ContractParameterType, Neo.BlockchainToolkit.ContractDataDef.StructDef>;
 using ContractParameterType = Neo.SmartContract.ContractParameterType;
 using System;
+using Neo.BlockchainToolkit.ContractDataDef;
 
 namespace test.bctklib
 {
@@ -32,35 +33,23 @@ namespace test.bctklib
                 },
             });
 
-            var us = ContractDataDef.ParseStructs(doc.RootElement);
-            var bs = ContractDataDef.BindStructs(us);
+            var us = StructDef.ParseStructProp(doc.RootElement);
+            var bs = StructDef.BindStructs(us);
 
-            var expect1 = new StructDef
-            {
-                Name = "TokenState",
-                Fields = new[] {
+            var expect1 = new StructDef("TokenState", new[] {
                     ("Owner", Field(ContractParameterType.Hash160)),
                     ("Name", Field(ContractParameterType.String)),
                     ("Description", Field(ContractParameterType.String)),
                     ("Image", Field(ContractParameterType.String)),
-                }
-            };
+                });
 
-            var expect2 = new StructDef
-            {
-                Name = "SomeTest",
-                Fields = new[] {
+            var expect2 = new StructDef("SomeTest", new[] {
                     ("foo", Field(ContractParameterType.Hash160)),
                     ("ts", FieldType.FromT1(expect1)),
-                }
-            };
+                });
 
-            var actual1 = bs.Single(s => s.Name == expect1.Name);
-            var actual2 = bs.Single(s => s.Name == expect2.Name);
-
-            bs.Should().HaveCount(2);
-            actual1.Should().BeEquivalentTo(expect1);
-            actual2.Should().BeEquivalentTo(expect2);
+            bs.Should().HaveCount(2)
+                .And.Contain(new [] { expect1, expect2 });
         }
 
         [Fact]
@@ -75,8 +64,8 @@ namespace test.bctklib
                 },
             });
 
-            var us = ContractDataDef.ParseStructs(doc.RootElement);
-            Assert.Throws<Exception>(() => ContractDataDef.BindStructs(us));
+            var us = StructDef.ParseStructProp(doc.RootElement);
+            Assert.Throws<Exception>(() => StructDef.BindStructs(us));
         }
 
         static FieldType Field(Neo.SmartContract.ContractParameterType param) => FieldType.FromT0(param);
